Split into sn0int and sn0int-std
