+++ Cargo.toml
@@ -14,13 +14,15 @@ travis-ci = { repository = "kpcyrd/sn0int" }
 
 [workspace]
 members = ["sn0int-registry/sn0int-common",
-           "sn0int-registry"]
+           "sn0int-registry",
+           "sn0int-std"]
 
 [features]
 sqlite-bundled = ["libsqlite3-sys/bundled"]
 
 [dependencies]
 sn0int-common = { version="0.9.0", path="sn0int-registry/sn0int-common" }
+sn0int-std = { version="0.16.0", path="sn0int-std" }
 rustyline = "5.0"
 log = "0.4"
 env_logger = "0.7"
@@ -31,7 +33,6 @@ rand = "0.7"
 colored = "1.6"
 lazy_static = "1.0"
 shellwords = "1.0"
-publicsuffix = { version="1.5", default-features=false }
 diesel = { version = "1.0.0", features = ["sqlite", "chrono"] }
 diesel_migrations = { version = "1.3.0", features = ["sqlite"] }
 libsqlite3-sys = "0.16.0"
@@ -41,14 +42,8 @@ url = "2.0"
 percent-encoding = "2.1"
 #chrootable-https = { path = "../chrootable-https" }
 chrootable-https = "0.14"
-rustls = { version="0.16", features=["dangerous_configuration"] }
-webpki = "0.21"
-webpki-roots = "0.18"
-ct-logs = "0.6"
-pem = "0.7"
 base64 = "0.11"
 data-encoding = "2.1.2"
-kuchiki = "0.7.2"
 serde_urlencoded = "0.6"
 serde = "1.0"
 serde_derive = "1.0"
@@ -61,34 +56,22 @@ maplit = "1.0.1"
 sloppy-rfc4880 = "0.1.5"
 regex = "1.0"
 toml = "0.5"
-maxminddb = "0.13"
 threadpool = "1.7"
-x509-parser = "0.6.0"
-der-parser = "3.0"
 atty = "0.2"
-bufstream = "0.1.4"
-tokio = "0.1.14"
 semver = "0.9"
 bytes = "0.4"
-xml-rs = "0.8"
 bytesize = "1.0"
 ipnetwork = "0.15"
 strum = "0.17"
 strum_macros = "0.17"
-tungstenite = { version = "0.9", default-features = false }
-geo = "0.12"
 
 digest = "0.8.0"
-bs58 = "0.3"
-blake2 = "0.8.0"
 md-5 = "0.8.0"
 sha-1 = "0.8.1"
 sha2 = "0.8.0"
 sha3 = "0.8.0"
 hmac = "0.7"
 
-image = "0.22"
-kamadak-exif = "0.4"
 walkdir = "2.2"
 nude = "0.2"
