+++ sn0int-std/Cargo.toml
@@ -1,0 +1,49 @@
+[package]
+name = "sn0int-std"
+version = "0.16.0"
+authors = ["kpcyrd <git@rxv.cc>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+log = "0.4"
+failure = "0.1"
+hlua-badtouch = "0.4"
+tokio = "0.1.14"
+rand = "0.7"
+regex = "1.0"
+rustls = { version="0.16", features=["dangerous_configuration"] }
+serde = "1.0"
+serde_derive = "1.0"
+serde_json = "1.0"
+serde_urlencoded = "0.6"
+webpki = "0.21"
+webpki-roots = "0.18"
+ct-logs = "0.6"
+chrootable-https = "0.14"
+bufstream = "0.1.4"
+pem = "0.7"
+url = "2.0"
+tungstenite = { version = "0.9", default-features = false }
+kuchiki = "0.7.2"
+maxminddb = "0.13"
+x509-parser = "0.6.0"
+der-parser = "3.0"
+publicsuffix = { version="1.5", default-features=false }
+xml-rs = "0.8"
+geo = "0.12"
+bytes = "0.4"
+base64 = "0.11"
+chrono = { version = "0.4", features = ["serde"] }
+
+image = "0.22"
+kamadak-exif = "0.4"
+
+bs58 = "0.3"
+digest = "0.8.0"
+blake2 = "0.8.0"
+
+[dev-dependencies]
+env_logger = "0.7"
+maplit = "1.0.1"
