+++ sn0int-std/src/lazy.rs
