+++ src/lib.rs
@@ -16,25 +16,25 @@ pub mod autonoscope;
 pub mod blobs;
 pub mod cmd;
 pub mod config;
-pub mod crt;
+use sn0int_std::crt;
 pub mod db;
 pub mod errors;
 pub mod engine;
 pub mod filters;
 pub mod fmt;
-pub mod geo;
-pub mod geoip;
-pub mod gfx;
-pub mod html;
-pub mod json;
+use sn0int_std::geo;
+pub use sn0int_std::geoip;
+use sn0int_std::gfx;
+use sn0int_std::html;
+use sn0int_std::json;
 pub mod keyring;
-pub mod lazy;
+use sn0int_std::lazy;
 pub mod migrations;
 pub mod models;
 pub mod paths;
-pub mod psl;
+pub use sn0int_std::psl;
 pub mod options;
-pub mod ratelimits;
+use sn0int_std::ratelimits;
 pub mod registry;
 pub mod repl;
 pub mod runtime;
@@ -42,17 +42,12 @@ pub mod sandbox;
 pub mod schema;
 pub mod ser;
 pub mod shell;
-pub mod sockets;
+use sn0int_std::sockets;
 pub mod term;
 pub mod update;
 pub mod utils;
-pub mod web;
-pub mod websockets;
+use sn0int_std::web;
+use sn0int_std::websockets;
 pub mod worker;
 pub mod workspaces;
-pub mod xml;
-
-#[cfg(test)]
-fn test_init() {
-    let _ = env_logger::builder().is_test(true).try_init();
-}
+use sn0int_std::xml;
